refactor: visaRegistration 및 applicants 페이지의 mock, style 파일 분리
