+++ src/features/applicants/ApplicantList/ContractModal/ModalText/index.style.ts
@@ -1,0 +1,13 @@
+export const titleStyle = {
+  fontWeight: 'bold',
+  marginBottom: '20px',
+};
+
+export const paragraphStyle = {
+  marginBottom: '20px',
+};
+
+export const headingStyle = {
+  ...titleStyle,
+  marginTop: '20px',
+};
