+++ src/features/applicants/ApplicantList/ContractModal/ModalText/index.tsx
@@ -1,25 +1,24 @@
 import { palettes } from '@/assets/styles/global/palettes';
+import { ForeignerData } from '@/types';
 import { Typo } from '@/components/common';
+import { headingStyle, paragraphStyle, titleStyle } from './index.style';
 
-interface ForeignerInfoProps {
-  foreignerIdNumber: string;
-  visaGenerateDate: string;
-}
+type Props = Pick<ForeignerData, 'foreignerIdNumber' | 'visaGenerateDate'>;
 
-export default function ModalText({ foreignerIdNumber, visaGenerateDate }: ForeignerInfoProps) {
+export default function ModalText({ foreignerIdNumber, visaGenerateDate }: Props) {
   return (
     <>
       <Typo element="h3" size="18px" style={titleStyle}>
         ✅ 고용주님께 드리는 주의사항
       </Typo>
-      <Typo element="p" size="14px" style={{ marginBottom: '20px' }}>
+      <Typo element="p" size="14px" style={paragraphStyle}>
         불법 체류자를 고용할 시 최대 <strong>200만원</strong>의 범칙금이 부과될 수 있습니다.
       </Typo>
       <hr />
-      <Typo element="h4" size="14px" style={{ ...titleStyle, marginTop: '20px' }}>
+      <Typo element="h4" size="14px" style={headingStyle}>
         지원자 정보
       </Typo>
-      <Typo element="p" size="14px" style={{ marginBottom: '20px' }}>
+      <Typo element="p" size="14px" style={paragraphStyle}>
         외국인 등록 번호 : <strong>{foreignerIdNumber}</strong>
         <br />
         비자 발급 일자 : <strong>{visaGenerateDate}</strong>
@@ -43,8 +42,3 @@ export default function ModalText({ foreignerIdNumber, visaGenerateDate }: Forei
     </>
   );
 }
-
-const titleStyle = {
-  fontWeight: 'bold',
-  marginBottom: '20px',
-};
