+++ src/features/applicants/ApplicantList/ContractModal/index.mock.ts
@@ -1,0 +1,5 @@
+export const foreigner = {
+  foreignerIdNumber: '123456-1234567',
+  visaGenerateDate: '2000-00-00',
+  visaExpiryDate: '2000-00-00',
+};
