+++ src/features/applicants/ApplicantList/ContractModal/index.style.ts
@@ -1,0 +1,15 @@
+import { palettes } from '@/assets/styles/global/palettes';
+import { css } from '@emotion/react';
+
+export const modalStyle = {
+  padding: '15px',
+};
+
+export const customButtonStyle = css`
+  background-color: ${palettes.blue};
+  color: ${palettes.white};
+`;
+
+export const buttonTextStyle = {
+  color: `${palettes.white}`,
+};
