+++ src/features/applicants/ApplicantList/ContractModal/index.tsx
@@ -1,13 +1,7 @@
 import { Button, Flex, Icon, Modal, Typo } from '@/components/common';
 import ModalText from './ModalText';
-import { css } from '@emotion/react';
-import { palettes } from '@/assets/styles/global/palettes';
-
-const initialData = {
-  foreignerIdNumber: '123456-1234567',
-  visaGenerateDate: '2000-00-00',
-  visaExpiryDate: '2000-00-00',
-};
+import { foreigner } from './index.mock';
+import { buttonTextStyle, customButtonStyle, modalStyle } from './index.style';
 
 interface ContractModalProps {
   isOpen: boolean;
@@ -20,17 +14,14 @@ export default function ContractModal({ isOpen, onClose }: ContractModalProps) {
       {isOpen && (
         <Modal
           textChildren={
-            <ModalText
-              foreignerIdNumber={initialData.foreignerIdNumber}
-              visaGenerateDate={initialData.visaGenerateDate}
-            />
+            <ModalText foreignerIdNumber={foreigner.foreignerIdNumber} visaGenerateDate={foreigner.visaGenerateDate} />
           }
           buttonChildren={
             <Flex justifyContent="space-between">
               <Button onClick={onClose}>취소</Button>
               <Button onClick={onClose} css={customButtonStyle}>
                 <Flex gap={{ x: '15px' }}>
-                  <Typo size="16px" style={{ color: `${palettes.white}` }}>
+                  <Typo size="16px" style={buttonTextStyle}>
                     확인하였습니다.
                   </Typo>
                   <Icon.Arrow.RightWhite />
@@ -39,14 +30,9 @@ export default function ContractModal({ isOpen, onClose }: ContractModalProps) {
             </Flex>
           }
           onClose={onClose}
-          style={{ padding: '15px' }}
+          style={modalStyle}
         />
       )}
     </>
   );
 }
-
-const customButtonStyle = css`
-  background-color: ${palettes.blue};
-  color: ${palettes.white};
-`;
