+++ src/features/visaRegistration/index.tsx
@@ -1,7 +1,6 @@
-import { responsiveStyle } from '@utils/responsive';
 import { Button, Flex, Input, Modal } from '@/components/common';
-import styled from '@emotion/styled';
 import { ChangeEvent, useState } from 'react';
+import { buttonStyle, ErrorMessage, Form, inputStyle } from './index.style';
 
 export default function VisaRegistrationForm() {
   const [foreignerNumber, setForeignerNumber] = useState('');
@@ -76,40 +75,3 @@ export default function VisaRegistrationForm() {
     </>
   );
 }
-
-const Form = styled.form`
-  display: flex;
-  flex-direction: column;
-  gap: 40px;
-  font-size: 16px;
-  font-weight: bold;
-  width: 100%;
-  max-width: 700px;
-  margin: 0 auto;
-
-  ${responsiveStyle({
-    tablet: {
-      gap: '20px',
-      padding: '0 20px',
-    },
-    mobile: {
-      padding: '0 15px',
-    },
-  })}
-`;
-
-const inputStyle = {
-  padding: '15px 20px',
-  width: '100%',
-};
-
-const buttonStyle = {
-  backgroundColor: '#0A65CC',
-  color: '#fff',
-  borderRadius: '4px',
-};
-
-const ErrorMessage = styled.div`
-  color: red;
-  font-size: 13px;
-`;
