+++ src/pages/applicants/index.mock.ts
@@ -1,0 +1,28 @@
+import CompanyImage from '@assets/images/coupang.png';
+
+export const recruitment = {
+  image: CompanyImage,
+  companyName: '쿠팡 유성점',
+  koreanTitle: '쿠팡 유성점에서 아르바이트 모집합니다.',
+  area: '대전 유성구',
+  salary: 50000000,
+};
+
+export const applicantList = [
+  {
+    userId: 1,
+    name: '이름1',
+    resumeId: 1,
+    applyId: 1,
+    applicantNation: '베트남',
+    korean: '중급',
+  },
+  {
+    userId: 2,
+    name: '이름2',
+    resumeId: 2,
+    applyId: 2,
+    applicantNation: '베트남',
+    korean: '고급',
+  },
+];
